Sort observations by time
